+++ api/ring-camera.ts
@@ -141,7 +141,8 @@ export class RingCamera extends Subscribed {
     private initialData: CameraData,
     public isDoorbot: boolean,
     private restClient: RingRestClient,
-    private avoidSnapshotBatteryDrain: boolean
+    private avoidSnapshotBatteryDrain: boolean,
+    private streamingConnectionOptions: StreamingConnectionOptions
   ) {
     super()
 
@@ -338,10 +339,14 @@ export class RingCamera extends Subscribed {
     return response.device_health
   }
 
-  private async createStreamingConnection(options: StreamingConnectionOptions) {
+  private async createStreamingConnection(options?: StreamingConnectionOptions) {
     if (this.isRingEdgeEnabled) {
       const auth = await this.restClient.getCurrentAuth()
-      return new RingEdgeConnection(auth.access_token, this, options)
+      return new RingEdgeConnection(
+        auth.access_token,
+        this,
+        options || this.streamingConnectionOptions
+      )
     }
 
     const liveCall = await this.restClient
@@ -359,10 +364,14 @@ export class RingCamera extends Subscribed {
         throw e
       })
 
-    return new WebrtcConnection(liveCall.data.session_id, this, options)
+    return new WebrtcConnection(
+      liveCall.data.session_id,
+      this,
+      options || this.streamingConnectionOptions
+    )
   }
 
-  async startLiveCall(options: StreamingConnectionOptions = {}) {
+  async startLiveCall(options?: StreamingConnectionOptions) {
     const connection = await this.createStreamingConnection(options)
     return new StreamingSession(this, connection)
   }
